requirements.txt
Django==1.5.1
GitPython==0.3.2.RC1
MySQL-python==1.2.4
argparse==1.2.1
async==0.6.1
distribute==0.6.34
gitdb==0.5.4
smmap==0.8.2
wsgiref==0.1.2
